requirements.txt
flask
bootstrap-flask
flask_login
flask_bootstrap
werkzeug
flask_sqlalchemy
flask_mysqldb
wtforms
wtforms
flask_wtf
flask_migrate
pymysql
wtforms
werkzeug
flask_login
blueprint
email_validator
passlib
flask_bcrypt
Pillow